requirements.txt
mcp==1.9.0
patch-ng==1.18.1
python-dotenv==1.1.0
python-dateutil==2.9.0.post0
unidiff==0.7.5
openai==1.79.0
sentence-transformers==4.1.0
tiktoken==0.9.0
pyyaml
pytest
